pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-gerard"

version = "0.1.0"
description = "Personal MCP toolkit for Gerard McCaul"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Gerard McCaul"}
]
dependencies = [
    "mcp[cli]>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "google-api-python-client>=2.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=0.5.0",
    "google-genai>=1.24.0",
    "openai>=1.0.0",
    "anthropic>=0.21.3",
    "mistralai>=1.9.0",
    "Pillow>=10.0.0",
    "httpx>=0.25.0",
    "packaging>=21.0",
    "PyYAML>=6.0.0",
    "click>=8.0.0",
    "html2text>=2020.1.16",
    "beautifulsoup4>=4.12.0",
    "markdownify>=0.11.0",
    "ftfy>=6.0.0",
    "python-dateutil>=2.8.0",
    "dateparser>=1.2.0",
    "pendulum>=3.0.0",
    "lxml>=4.9.0",
    "rapidfuzz>=3.0.0",
    "matplotlib>=3.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-vcr>=1.0.0",
    "pytest-xdist>=3.0.0",
    "vcrpy>=4.0.0",
    "nest_asyncio>=1.6.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0.0",
    "tomli>=2.0.0",
]

[project.scripts]
# Inherited tools from mcp-handley-lab
mcp-llm = "mcp_gerard.llm.tool:mcp.run"
mcp-llm-embeddings = "mcp_gerard.llm.embeddings.tool:mcp.run"
mcp-code2prompt = "mcp_gerard.code2prompt.tool:mcp.run"
mcp-arxiv = "mcp_gerard.arxiv.tool:mcp.run"
mcp-google-calendar = "mcp_gerard.google_calendar.tool:mcp.run"
mcp-loop = "mcp_gerard.loop.tool:mcp.run"
mcp-word = "mcp_gerard.microsoft.word.tool:mcp.run"
# Personal tools
mcp-blog = "mcp_gerard.blog:main"
mcp-overleaf = "mcp_gerard.overleaf:main"
mcp-vault = "mcp_gerard.vault:main"
mcp-projects = "mcp_gerard.projects:main"


[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
mcp_gerard = ["py.typed", "llm/providers/*/models.yaml", "tool_schemas.json", "_shared/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "unit: marks tests as unit tests (select with '-m unit')",
    "vcr: marks tests that use VCR.py for HTTP recording/playback",
    "slow: marks tests as slow running (deselect with '-m \"not slow\"')",
    "live: marks tests that require live API calls (deselect with '-m \"not live\"')",
]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501", "B008"]  # Line too long, function call in default arguments (intentional for FastMCP Field patterns)

[tool.ruff.lint.per-file-ignores]
"src/mcp_gerard/google_calendar/tool.py" = ["N815"]  # Allow mixedCase for Google Calendar API field names
"src/mcp_gerard/microsoft/opc/*.py" = ["N802", "N803", "N806"]  # OPC uses rId naming convention
"src/mcp_gerard/microsoft/word/*.py" = ["N802", "N803", "N806", "E402", "SIM102", "SIM105", "SIM108", "SIM117", "B007", "B904"]  # OOXML naming conventions
"src/mcp_gerard/microsoft/excel/*.py" = ["N802", "N803", "N806"]  # SpreadsheetML uses OOXML naming conventions
"src/mcp_gerard/microsoft/powerpoint/*.py" = ["N802", "N803", "N806"]  # PresentationML uses OOXML naming conventions
"tests/word/*.py" = ["N806"]  # Allow OOXML element naming in tests
"tests/excel/*.py" = ["N806"]  # Allow OOXML element naming in tests
"tests/powerpoint/*.py" = ["N806"]  # Allow OOXML element naming in tests

[tool.coverage.run]
omit = [
    "src/mcp_gerard/__main__.py",
    "tests/*",
    "venv/*",
]
